+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/ForgotPassword/Command.cs
@@ -1,0 +1,3 @@
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.ForgotPassword;
+
+public sealed record Command(string Email);