+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/ForgotPassword/Validator.cs
@@ -1,0 +1,13 @@
+using FluentValidation;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.ForgotPassword;
+
+public sealed class Validator : AbstractValidator<Command>
+{
+    public Validator()
+    {
+        RuleFor(x => x.Email)
+            .NotEmpty()
+            .EmailAddress();
+    }
+}