+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/ResetPassword/Command.cs
@@ -1,0 +1,3 @@
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.ResetPassword;
+
+public sealed record Command(string Token, string NewPassword);