+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/ResetPassword/Validator.cs
@@ -1,0 +1,15 @@
+using FluentValidation;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.ResetPassword;
+
+public sealed class Validator : AbstractValidator<Command>
+{
+    public Validator()
+    {
+        RuleFor(x => x.Token)
+            .NotEmpty();
+
+        RuleFor(x => x.NewPassword)
+            .NotEmpty();
+    }
+}